amount fields increased from 4 byte ints to 8 byte signed integers

+++ backend/migrations/oft_postgres.sql
@@ -25,7 +25,7 @@ CREATE TABLE accounts (
     user_id INTEGER NOT NULL,
     account_name TEXT NOT NULL,
     currency TEXT NOT NULL DEFAULT 'USD',
-    checkpoint_balance INTEGER NOT NULL,
+    checkpoint_balance BIGINT NOT NULL,
     checkpoint_timestamp TIMESTAMP NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (user_id) REFERENCES users(user_id) ON DELETE CASCADE
@@ -35,7 +35,7 @@ CREATE TABLE transacts (
     trans_id SERIAL PRIMARY KEY,
     account_id INTEGER NOT NULL,
     occurred_at TIMESTAMP NOT NULL,
-    amount_cents INTEGER NOT NULL CHECK (amount_cents >= 0),
+    amount_cents BIGINT NOT NULL CHECK (amount_cents >= 0),
     direction TEXT NOT NULL CHECK (direction IN ('credit','debit')),
     trans_status TEXT NOT NULL DEFAULT 'posted' CHECK (trans_status IN ('posted','deleted')),
     notes TEXT NOT NULL,
