+++ backend/src/app/db/models/accounts.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from typing import Optional
-from sqlalchemy import Integer, String, DateTime, ForeignKey, text
+from sqlalchemy import Integer, String, DateTime, ForeignKey, BigInteger, text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from . import Base
 
@@ -14,7 +14,7 @@ class Account(Base):
     currency:   Mapped[str] = mapped_column(String, nullable=False, server_default=text("'USD'"))
 
     # checkpoint value + timestamp (TEXT in SQL, mapped as datetime here)
-    checkpoint_balance: Mapped[int] = mapped_column(Integer, nullable=False)
+    checkpoint_balance: Mapped[int] = mapped_column(BigInteger, nullable=False)
     checkpoint_timestamp:            Mapped[datetime] = mapped_column(DateTime, nullable=False)
 
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=text("CURRENT_TIMESTAMP"), nullable=False)
