+++ backend/src/app/db/models/transacts.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Integer, String, DateTime, ForeignKey, CheckConstraint, text
+from sqlalchemy import Integer, String, DateTime, ForeignKey, BigInteger, CheckConstraint, text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from . import Base
 
@@ -10,7 +10,7 @@ class Transact(Base):
     account_id: Mapped[int] = mapped_column(ForeignKey("accounts.account_id", ondelete="CASCADE"), nullable=False)
 
     occurred_at:  Mapped[datetime] = mapped_column(DateTime, nullable=False)
-    amount_cents: Mapped[int] = mapped_column(Integer, nullable=False)  # positive integer cents
+    amount_cents: Mapped[int] = mapped_column(BigInteger, nullable=False)  # positive integer cents
     direction:    Mapped[str] = mapped_column(String, nullable=False)   # 'credit' | 'debit'
     trans_status:       Mapped[str] = mapped_column(String, nullable=False, server_default=text("'posted'"))  # 'posted' | 'deleted'
     notes:        Mapped[str] = mapped_column(String, nullable=False)
